requirements.txt
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
python-decouple==3.3
Flask-Migrate==4.0.0
psycopg2==2.9.5